refactor(user-service): User lib/router instead of mux
G

Remove user-service binary
